// src/app/models/profile.ts
export class Profile {
    id: string;
    name: string;
    lastName: string;
    ci: string;
    address: string;
    telephone: string;
    cellphone: string;
    country: string;
    city: string;
    email: string;

    constructor() {
        this.id = "";
        this.name = "";
        this.lastName = "";
        this.ci = '';
        this.address = "";
        this.telephone = "";
        this.cellphone = "";
        this.country = "";
        this.city = "";
        this.email = "";
    }
}